Move readability methods into LoggingDLL (ReadabilityExtension)
Update tests accordingly
