+++ Logging/anvil-csharp-logging.dll
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:04849f45fba2855d3d9e94fff164ed1166dc9b4163a74fdae0d73b933ae10b17
-size 13824
+oid sha256:b2029329a71b51c8b91e9341d642a03a2a22ca57596d6fee17602a4d48001f8d
+size 14848
